refactor: fix xo issues
