+++ src/constants.js
@@ -1,0 +1,2 @@
+export const SOURCE_NAME = `GraphCMS`;
+export const DEBUG_MODE = process.env.DEBUG_MODE || false;
