+++ src/gatsby-node.js
@@ -1,16 +1,12 @@
 import {GraphQLClient} from 'graphql-request';
-import {forEach, forEachObjIndexed} from 'ramda';
-import crypto from 'crypto';
-import {extractTypeName} from './util';
+import {forEachObjIndexed} from 'ramda';
+import {createNodes} from './util';
+import {DEBUG_MODE} from './constants';
 import {
   faultyKeywords,
   keywordsError,
   checkForFaultyFields
 } from './faulty-keywords';
-
-const SOURCE_NAME = `GraphCMS`;
-
-const DEBUG_MODE = process.env.DEBUG_MODE || false;
 
 exports.sourceNodes = async (
   {boundActionCreators, reporter},
@@ -40,33 +36,4 @@ exports.sourceNodes = async (
   } else {
     reporter.panic(`gatsby-source-graphcms: you need to provide a GraphQL query in the plugin 'query' parameter`);
   }
-};
-
-const createNodes = (createNode, reporter) => (value, key) => {
-  forEach(queryResultNode => {
-    const {id, ...fields} = queryResultNode;
-    const jsonNode = JSON.stringify(queryResultNode);
-    const gatsbyNode = {
-      id,
-      ...fields,
-      parent: `${SOURCE_NAME}_${key}`,
-      children: [],
-      internal: {
-        type: extractTypeName(key),
-        content: jsonNode,
-        contentDigest: crypto.createHash(`md5`).update(jsonNode).digest(`hex`)
-      }
-    };
-
-    if (DEBUG_MODE) {
-      const jsonFields = JSON.stringify(fields);
-      const jsonGatsbyNode = JSON.stringify(gatsbyNode);
-      reporter.info(`  processing node: ${jsonNode}`);
-      reporter.info(`    node id ${id}`);
-      reporter.info(`    node fields: ${jsonFields}`);
-      reporter.info(`    gatsby node: ${jsonGatsbyNode}`);
-    }
-
-    createNode(gatsbyNode);
-  }, value);
 };
